+++ src/main/java/domain/board/Position.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 import java.util.Objects;
