+++ src/main/java/domain/piece/Pawn.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.board.Position;
 
 public class Pawn extends Piece {
