+++ src/main/java/domain/piece/Piece.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.board.Position;
 
 public abstract class Piece {
