+++ src/main/java/domain/piece/Queen.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.board.Position;
 
 public class Queen extends Piece {
