+++ src/main/java/view/OutputView.java
@@ -1,8 +1,8 @@
 package view;
 
-import domain.File;
-import domain.Position;
-import domain.Rank;
+import domain.board.File;
+import domain.board.Position;
+import domain.board.Rank;
 import domain.piece.Bishop;
 import domain.piece.King;
 import domain.piece.Knight;
