+++ src/test/java/domain/board/FileTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
