+++ src/test/java/domain/piece/RookTest.java
@@ -1,8 +1,8 @@
 package domain.piece;
 
-import domain.File;
-import domain.Position;
-import domain.Rank;
+import domain.board.File;
+import domain.board.Position;
+import domain.board.Rank;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
